move folders backend, client, electron into "src" folder of the main app
