+++ package.json
@@ -4,7 +4,7 @@
   "description": "A table tennis software for organizing tournaments. Based on the rules of the bavarian TT Race Tournament.",
   "version": "1.0.0",
   "homepage": "./",
-  "main": "./electron/main.js",
+  "main": "./src/electron/main.js",
   "license": "MIT",
   "contributors": [
     "Felix Breitenbach",
@@ -27,13 +27,13 @@
     "start-dev-electron-mac": "cross-env NODE_ENV=dev electron .",
     "start-dev-electron": "cross-env NODE_ENV=dev nodemon --exec \"\"electron .\"\"",
     "start-react-app": "cross-env PORT=3000 react-scripts start",
-    "start-react-client": "cd client && cross-env PORT=3001 yarn start",
+    "start-react-client": "cd ./src/client && cross-env PORT=3001 yarn start",
     "start-simulation": "cd src/simulation && node testCompetition.js",
     "build": "yarn build-react-app && yarn build-react-client && yarn build-electron",
-    "build-electron": "cross-env-shell \"yarn run reset-electron && mkdir build/src && mkdir build/client && cp -r electron/. build/electron && cp -r src/shared/. build/src/shared && cp -r client/build/. build/client/build && cp -r backend/. build/backend && cp -r assets/. build/assets\"",
+    "build-electron": "cross-env-shell \"yarn run reset-electron && mkdir build/src && mkdir build/src/client && cp -r src/electron/. build/src/electron && cp -r src/shared/. build/src/shared && cp -r src/client/build/. build/src/client/build && cp -r src/backend/. build/src/backend && cp -r assets/. build/assets\"",
     "build-react-app": "react-scripts build",
-    "build-react-client": "cd client && yarn build",
-    "reset-electron": "cross-env \"rm -rf build/src && rm -rf build/electron && rm -rf build/client && rm -rf build/backend && rm -rf build/assets\"",
+    "build-react-client": "cd ./src/client && yarn build",
+    "reset-electron": "cross-env \"rm -rf build/src && rm -rf build/assets\"",
     "preelectron-pack": "yarn build",
     "electron-pack": "electron-builder build --win -c.extraMetadata.main=build/electron/main.js --publish never",
     "postinstall": "electron-rebuild electron-builder install-app-deps",
