+++ src/backend/persistance/competition-repository.js
