+++ src/client/public/logo512.png
