+++ src/client/src/index.js
