+++ src/client/src/react/components/MainScore/MainScore.js
