+++ src/client/src/react/components/MainScore/MainScore.test.js
