+++ src/client/src/react/components/MainScore/main-score-state-func.js
