+++ src/client/src/react/components/Player/Player.js
