+++ src/client/src/react/components/PlayerScore/PlayerScore.js
