+++ src/electron/config.js
@@ -4,7 +4,7 @@ const isDev = require('electron-is-dev');
 const url = require('url');
 const path = require('path');
 
-const network = require('../backend/utils/network-helper');
+const network = require('../utils/network-helper');
 
 // server config
 const SERVER_HOST = isDev ? 'localhost' : network.getIpAddress();
